sample naming tweaks
