+++ OutboxKit.sln
@@ -9,11 +9,11 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core", "src\Core\Core.cspro
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "samples", "samples", "{2C240B41-C6C9-4921-9139-ECB2AA468316}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EfMySqlSample", "samples\EfMySqlSample\EfMySqlSample.csproj", "{C479A883-CE60-4521-A5D5-725C5704D736}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MySqlEfPollingSample", "samples\MySqlEfPollingSample\MySqlEfPollingSample.csproj", "{C479A883-CE60-4521-A5D5-725C5704D736}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MySql", "src\MySql\MySql.csproj", "{16BB39E2-046B-4390-8375-37BCC15F81EF}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MultiTenantEfMySqlSample", "samples\MultiTenantEfMySqlSample\MultiTenantEfMySqlSample.csproj", "{4C7A81B7-D08C-4810-8C00-9C857C8C53E5}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MySqlEfMultiDbPollingSample", "samples\MySqlEfMultiDbPollingSample\MySqlEfMultiDbPollingSample.csproj", "{4C7A81B7-D08C-4810-8C00-9C857C8C53E5}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MongoDb", "src\MongoDb\MongoDb.csproj", "{87A70C79-3D50-45D0-BF2D-113262F92C94}"
 EndProject
