+++ samples/MySqlEfMultiDbPollingSample/Ef.cs
@@ -3,11 +3,22 @@ using Microsoft.EntityFrameworkCore.Diagnostics;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 using YakShaveFx.OutboxKit.Core.Polling;
 
-namespace EfMySqlSample;
+namespace MySqlEfMultiDbPollingSample;
 
-public sealed class SampleContext(DbContextOptions<SampleContext> options) : DbContext(options)
+public sealed class SampleContext(
+    DbContextOptions<SampleContext> options,
+    ITenantProvider tenantProvider,
+    ConnectionStringProvider connectionStringProvider) : DbContext(options)
 {
+    private static readonly MySqlServerVersion MySqlVersion = new(new Version(8, 0));
+
     public DbSet<OutboxMessage> OutboxMessages => Set<OutboxMessage>();
+
+    protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
+    {
+        optionsBuilder.UseMySql(connectionStringProvider.GetConnectionString(tenantProvider.Tenant), MySqlVersion);
+        base.OnConfiguring(optionsBuilder);
+    }
 
     protected override void OnModelCreating(ModelBuilder modelBuilder)
         => modelBuilder.ApplyConfigurationsFromAssembly(GetType().Assembly);
@@ -27,27 +38,34 @@ public sealed class OutboxMessageConfiguration : IEntityTypeConfiguration<Outbox
 {
     public void Configure(EntityTypeBuilder<OutboxMessage> builder)
     {
+        builder.ToTable("outbox_messages");
         builder.HasKey(e => e.Id);
-        builder.Property(e => e.Id);
-        builder.Property(e => e.Target).HasMaxLength(128);
-        builder.Property(e => e.Type).HasMaxLength(128);
-        builder.Property(e => e.Payload);
-        builder.Property(e => e.CreatedAt);
-        builder.Property(e => e.ObservabilityContext);
+        builder.Property(e => e.Id).HasColumnName("id");
+        builder.Property(e => e.Target).HasColumnName("target").HasMaxLength(128);
+        builder.Property(e => e.Type).HasColumnName("type").HasMaxLength(128);
+        builder.Property(e => e.Payload).HasColumnName("payload");
+        builder.Property(e => e.CreatedAt).HasColumnName("created_at");
+        builder.Property(e => e.ObservabilityContext).HasColumnName("observability_context");
     }
 }
 
-public sealed class DbSetupHostedService(IServiceProvider serviceProvider) : BackgroundService
+public sealed class DbSetupHostedService(IServiceProvider serviceProvider, TenantList tenantList) : BackgroundService
 {
     protected override async Task ExecuteAsync(CancellationToken stoppingToken)
     {
-        using var scope = serviceProvider.CreateScope();
-        var context = scope.ServiceProvider.GetRequiredService<SampleContext>();
-        await context.Database.EnsureCreatedAsync(stoppingToken);
+        foreach (var tenant in tenantList)
+        {
+            using var scope = serviceProvider.CreateScope();
+            var context = scope.ServiceProvider.GetRequiredService<SampleContext>();
+            var tenantProvider = scope.ServiceProvider.GetRequiredService<TenantProvider>();
+            tenantProvider.SetTenant(tenant);
+            await context.Database.EnsureCreatedAsync(stoppingToken);
+        }
     }
 }
 
-public sealed class OutboxInterceptor(IOutboxTrigger trigger) : SaveChangesInterceptor
+public sealed class OutboxInterceptor(IKeyedOutboxTrigger trigger, ITenantProvider tenantProvider)
+    : SaveChangesInterceptor
 {
     private bool _hasOutboxMessages;
 
@@ -73,7 +91,7 @@ public sealed class OutboxInterceptor(IOutboxTrigger trigger) : SaveChangesInter
         {
             // this isn't mandatory, but if we don't trigger it after adding messages to the outbox, they will only be published on the next polling iteration
             // if waiting for polling iterations is acceptable, then don't call this:  code gets simpler and the db is less loaded
-            trigger.OnNewMessages();
+            trigger.OnNewMessages(tenantProvider.Tenant);
         }
 
         return await base.SavedChangesAsync(eventData, result, cancellationToken);
