+++ samples/MySqlEfMultiDbPollingSample/MySqlEfMultiDbPollingSample.csproj
@@ -4,6 +4,7 @@
         <TargetFramework>net8.0</TargetFramework>
         <Nullable>enable</Nullable>
         <ImplicitUsings>enable</ImplicitUsings>
+        <RootNamespace>MySqlEfMultiDbPollingSample</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
