+++ samples/MySqlEfMultiDbPollingSample/Properties/launchSettings.json
