+++ samples/MySqlEfPollingSample/MySqlEfPollingSample.csproj
@@ -4,6 +4,7 @@
     <TargetFramework>net8.0</TargetFramework>
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
+    <RootNamespace>MySqlEfPollingSample</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
