+++ samples/MySqlEfPollingSample/Properties/launchSettings.json
