+++ samples/MySqlEfPollingSample/appsettings.Development.json
