+++ samples/MySqlEfPollingSample/appsettings.json
