+++ samples/MySqlEfPollingSample/requests.http
